src/requirements.txt
python-dotenv=1.0.1
streamlit=1.33.0
google-generativeai=0.4.1